Introduce DistributedServerServiceSelectionAlgorithm.
This is now the default algorithm, and will attempt to distribute Server Instances across servers to the best of it's ability.  This will help in hardware / host redundancy at runtime.
Also, minor tweak to missing services, it will attempt to fill in from bottom-up now.  This means that the serviceInstances section in the application is now designed like an architecture diagram, with the most important on top.